fixed ogDescription
