+++ package.json
@@ -71,6 +71,7 @@
     "highcharts": "^11.4.1",
     "highcharts-react-official": "^3.2.1",
     "hono": "^4",
+    "html-to-text": "^9.0.5",
     "html2canvas": "^1.4.1",
     "js-cookie": "^3.0.5",
     "konva": "^9.3.14",
@@ -103,6 +104,7 @@
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/dompurify": "^3.0.5",
+    "@types/html-to-text": "^9.0.4",
     "@types/node": "^17.0.31",
     "@types/quill": "^2.0.14",
     "@types/react": "^18.3.1",
