+++ src/components/common/text-editor/index.tsx
@@ -1,9 +1,9 @@
 'use client'
 
 import { Box } from '@chakra-ui/react'
-import DOMPurify from 'dompurify'
 import dynamic from 'next/dynamic'
 import 'react-quill/dist/quill.snow.css'
+import { isFormattedText, formatTextWithLinks } from '@/utils/html-utils'
 
 const ReactQuill = dynamic(() => import('react-quill'), { ssr: false })
 
@@ -14,63 +14,6 @@ export type TextEditorProps = Readonly<{
   style?: React.CSSProperties
   className?: string
 }>
-
-const linkify = (text: string) => {
-  const urlRegex = /(https?:\/\/[^\s]+)/g
-  const escapeHtml = (unsafe: string) => {
-    return unsafe
-      .replace(/&/g, '&amp;')
-      .replace(/</g, '&lt;')
-      .replace(/>/g, '&gt;')
-      .replace(/"/g, '&quot;')
-      .replace(/'/g, '&#039;')
-  }
-  const isValidUrl = (url: string) => {
-    try {
-      new URL(url)
-      return true
-    } catch {
-      return false
-    }
-  }
-  return text.replace(urlRegex, (url) =>
-    isValidUrl(url)
-      ? `<a href="${escapeHtml(url)}" target="_blank" rel="noopener noreferrer">${escapeHtml(
-          url
-        )}</a>`
-      : url
-  )
-}
-
-const isFormattedText = (text: string): boolean => {
-  if (!text || typeof text !== 'string') return false
-  const htmlPatterns = [
-    /<[a-z][\s\S]*?>/i, // General HTML tags
-    /&(?:#\d+|\w+);/, // HTML entities
-    /<(p|div|span|a|strong|em|u|s|ul|ol|li|br|img|h[1-6]|blockquote|pre|code)\b/i, // Quill-related tags
-  ]
-
-  // Check if text matches any of the patterns
-  const containsHtmlTags = htmlPatterns.some((regex) => regex.test(text))
-
-  // Check if stripping HTML produces a different result
-  const strippedText = stripHTML(text)
-  const isPlainText = strippedText === text
-
-  // Determine if the text is formatted
-  return containsHtmlTags || !isPlainText
-}
-
-// Helper function to strip HTML
-const stripHTML = (html: string): string => {
-  const tmp = document.createElement('div')
-  try {
-    tmp.innerHTML = DOMPurify.sanitize(html) // Sanitize input for safety
-    return tmp.textContent || tmp.innerText || ''
-  } finally {
-    tmp.remove()
-  }
-}
 
 export default function TextEditor({
   value = '',
@@ -86,29 +29,7 @@ export default function TextEditor({
     if (!value) return ''
 
     if (readOnly) {
-      if (isHtml) {
-        const div = document.createElement('div')
-        div.innerHTML = DOMPurify.sanitize(value)
-
-        const walkTextNodes = (node: Node) => {
-          if (node.nodeType === Node.TEXT_NODE) {
-            const newText = linkify(node.textContent || '')
-            if (newText !== node.textContent) {
-              const span = document.createElement('span')
-              span.innerHTML = newText
-              node.parentNode?.replaceChild(span, node)
-            }
-          } else {
-            if (node.nodeName !== 'A') {
-              node.childNodes.forEach(walkTextNodes)
-            }
-          }
-        }
-
-        walkTextNodes(div)
-        return div.innerHTML
-      }
-      return linkify(value)
+      return formatTextWithLinks(value, isHtml)
     }
     return value
   }
